+++ lib/RegistrationLoginOption.php
@@ -28,22 +28,33 @@ namespace OCA\Registration;
 use OCP\Authentication\IAlternativeLogin;
 use OCP\IL10N;
 use OCP\IURLGenerator;
+use OCP\IConfig;
 
 class RegistrationLoginOption implements IAlternativeLogin {
 
-	public function __construct(protected IURLGenerator $url, protected IL10N $l, protected \OC_Defaults $theming) {
+	public function __construct(protected IConfig $config, protected IURLGenerator $url, protected IL10N $l, protected \OC_Defaults $theming) {
+		$this->config = $config;
 	}
 
 	public function getLabel(): string {
-		return $this->l->t('Register');
+		// Decide if login page will display registration link (thisiszeev)
+		if ($this->config->getAppValue('registration', 'login_button_hide', 'no') === 'no') {
+			return $this->l->t('Register');
+		}
 	}
 
 	public function getLink(): string {
-		return $this->url->linkToRoute('registration.register.showEmailForm');
+		// Decide if login page will display registration link (thisiszeev)
+		if ($this->config->getAppValue('registration', 'login_button_hide', 'no') === 'no') {
+			return $this->url->linkToRoute('registration.register.showEmailForm');
+		}
 	}
 
 	public function getClass(): string {
-		return 'register-button';
+		// Decide if login page will display registration link (thisiszeev)
+		if ($this->config->getAppValue('registration', 'login_button_hide', 'no') === 'no') {
+			return 'register-button';
+		}
 	}
 
 	public function load(): void {